# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: G Creator Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /upload/media:
    post:
      operationId: uploadMedia
      tags: [upload]
      summary: Upload a public media file (photo, video, document)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/MediaUploadRequest"
      responses:
        "200":
          description: Public URL and storage key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaUploadResponse"
  /upload/private:
    post:
      operationId: uploadPrivate
      tags: [upload]
      summary: Upload a private document (ID verification, sensitive files)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/PrivateUploadRequest"
      responses:
        "200":
          description: Storage key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrivateUploadResponse"
  /upload/private-url:
    get:
      operationId: getPrivateDownloadUrl
      tags: [upload]
      summary: Get a time-limited download URL for a private file
      security:
        - bearerAuth: []
      parameters:
        - name: key
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Time-limited download URL
          content:
            application/json:
              schema:
                type: object
                required: [downloadUrl]
                properties:
                  downloadUrl:
                    type: string
...
  /mux/upload/{uploadId}:
    get:
      operationId: getMuxUploadStatus
      tags: [mux]
      summary: Poll status of a Mux direct upload
      security:
        - bearerAuth: []
      parameters:
        - name: uploadId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Upload status and playback ID when ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MuxUploadStatus"
  /mux/uploads/{uploadId}:
    delete:
      operationId: cancelMuxUpload
      tags: [mux]
      summary: Cancel an in-progress Mux upload and remove its pending record
      security:
        - bearerAuth: []
      parameters:
        - name: uploadId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Upload cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required: [ok]
        "404":
          description: Upload not found or not owned by caller
components:
  schemas:
...
    MediaUploadRequest:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
          description: File to upload
        folder:
          type: string
          description: "Destination folder (default: media)"
    MediaUploadResponse:
      type: object
      required: [publicUrl, key]
      properties:
        publicUrl:
          type: string
        key:
          type: string
    PrivateUploadRequest:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
          description: File to upload
        folder:
          type: string
          description: "Destination folder (default: docs)"
    PrivateUploadResponse:
      type: object
      required: [key]
      properties:
        key:
          type: string
...
    MuxUploadStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [waiting, asset_created, errored, cancelled]
        assetId:
          type: string
        playbackId:
          type: string